refactor: move functions to classes - Move "get", filter and display functions from index.js to dedicated classes.
